pyproject.toml
[project]
name = "synfact"
version = "0.1.0"
description = "Synthetic data generation engine for SynFact-L benchmark"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.0.0",
    "pydantic>=2.0.0",
    "datasets>=2.0.0",
    "rich>=13.0.0",
    "tenacity>=8.0.0",
    "python-dotenv>=1.2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
